Committing before closing test suite branch.
Testing suite is able to produce artificial data files in a robust manner so that I can now build the rest of the plotting, analysis and prediction code. Time to close this branch and start developing the plotting functions in conjunction with whatever additions need to be made to the FittedModel class to interface with the plotting functions.