Having finished reduced chi-squared build.
Pushing and merging this branch. Everything seems to be working but we are at the point where we need to working with much larger (but still test) datasets, so next task is to get things running on HTCondor. Pushing and merging this branch and will then make a branch for HTCondor.