Have found some speed.
Have reparameterised our linear, hierarchical multi-normal model as described in the Stan User's Guide. On the laptop a 40 day, 1000 observation per day test dataset takes less than an hour and a half to fit with no complaints from stan, so this is fine to proceed with. Pushing now so that I can pull to the IoA and run a full run there.