Have implemented model comparison.
Can estimate elpd for both models. Still need to change that function slightly to make it more efficient. Have also made it so that the data preparation of json files is either error-averaged or not depending on which model you're using. This will eventually be done away with once we confirm that the fast model is fine via elpd WAIC.