Committing to pull.
elpd_waic can be calculated and compared for both models and the results are promising. Committing to pull into main branch and start the next one, which will be a big one. Going to make a new module that will convert a year's worth of TROPOMI observations into a usable dataset.csv