Added plotables csv function
Created a single function to create a large csv file of all possible plotable quantities so far. This function will eventually be edited in the future to include data poor days. Next task is to use this csv in all plotting functions, and then remove unecessary functions that are no longer needed.